requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
distlib==0.3.8
Django==5.1.1
django-crispy-forms==2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2
filelock==3.16.1
platformdirs==4.3.6
PyJWT==2.9.0
pytz==2024.2
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
virtualenv==20.26.5